disp: msm: dp: Implement dsc slice selection logic based on sink caps
The dp sink explicitly states the numbers of
slices it can support. When calculating the
number of slices for dsc configuration, ensure
that the sink supports the calculated number of
slices.

If the sink does not support the calculated
number of slices, keep rechecking sink support
for the subsequent slice increment, until we
reach the highest number of slices possible
according to the dp spec. If no compatible
sink slice support is found, do not enable dsc.

Change-Id: I2db59b0f63a62576ca28276283b7c786f9fbaaf2
Signed-off-by: Fuad Hossain <fhossain@codeaurora.org>
Signed-off-by: Xiaowen Wu <wxiaowen@codeaurora.org>