drm: msm: sde: reserve LM in pairs for 4LM topologies
Rreserve LM in the sequence of primary-peer and then
primary-peer again until all required LM are reserved.

Searching sequence will not change for single/dual LM
reservation. For single LM reservation it will return
right after the first primary LM is found. For dual LM
reservation it will return after the first primary-peer
pair is found.

The logic can also work for triple/quad or any number
of LM reservetion and make sure that all the reserved
LMs are in pairs except the last one if total LM number
is odd.

This change also looses the DSPP check for 4LM topologies,
that if 4LM is needed and DSPP is not required, allow
DSPP-enabled LMs to be reserved.

Change-Id: Ia28bb64fedeb43430039775051943d751259a3d2
Signed-off-by: Amine Najahi <anajahi@codeaurora.org>
Signed-off-by: satbir singh <satbsing@codeaurora.org>
Signed-off-by: Xiaowen Wu <wxiaowen@codeaurora.org>