disp: msm: dp: fix gpio hpd detection for hpd glitch
Optimize the logic for gpio hpd detection to avoid missing hpd.

Currently when hpd is high and low in a short period of time,
the hpd low will not be detected as gpio detection is switched
to falling edge after hpd is already low.

The new change will always read the hpd state after edge detection
switch, to make sure the next hpd change can be detected.

Change-Id: I2f7c0823e6e37ff8d68257db9e2eae122a8e29a2
Signed-off-by: Xiaowen Wu <wxiaowen@codeaurora.org>