disp: msm: add get_tile_map support to sde connector
Add get_tile_map support to sde connection. This callback function
will be called when setup the tile mode and will return the mapping
of logical h_tile to physical port.

For DP, tile_h_loc from EDID will be used as the tile index of each
connector.

For DSI, qcom,mdss-dsi-panel-swap-intf in panel mode will be used
to indicate if h_tile needs to be swapped.

Change-Id: Id521b82ae5a33f0ebd2d278db44a0c05d6ec9bf6
Signed-off-by: Xiaowen Wu <wxiaowen@codeaurora.org>