disp: msm: shd: add option to use connector name as base display
Currently driver is using HW intf index in device tree to select
shared base display, which is less human readable for integrator
to configure. This change allows connector name as an alternative
option to be the base display identifier like DSI-1, DSI-2, DP-1,
etc, which is also aligned with multi-cards configuration.

Change-Id: I5b6c029484e927364e9a1c2fbf2f2acf3067fae9
Signed-off-by: Xiaowen Wu <wxiaowen@codeaurora.org>