disp: msm: shp: add enable_req option to handoff property
In the shared plane design, only the owner of the active shared
plane can handoff the plane to another compositor, which brings
difficulty when handoff is requested from the compositor that
owns the inactive shared plane.

Assume we have shared plane A and B, plane B is in active state and
plane A is disabled with possible_crtcs=0. When compositor tries to
activate plane A, it needs a way to notify the compositor of plane B
to handoff plane B first. Currently this is done through user defined
IPC.

One difficulty of using IPC is that usually compositor of plane A has
no knowledge of plane B when it requests to handoff plane B.
When there are multiple planes shared across different containers,
the owner may own multiple active shared planes and doesn't know
which plane to handoff from the IPC request.

This change provides a new method that instead of talking to the
owner of the active plane, request the owner of the inactive
plane to re-enable the plane.

With the change, when plane A set handoff to enable_req stte,
shared plane driver will generate handoff uevent with the id of
active plane B. Now the owner of plane B knows there is request
to handed off plane B. No additional IPC call is required.

Change-Id: I43ef8f6dce2da5f5fdb3dcdb8faa9fa6f50edb55
Signed-off-by: Xiaowen Wu <wxiaowen@codeaurora.org>