disp: msm-lease: allow lease card to open when card0 is opened
When lessee cards are opened, card0 should not be opened as master
because card0 has the access to all resources and is not aware that
some resources are leased to other cards. Currently driver will
reject the lessee card to open when card0 is already opened as
master.

Since drmOpen function always walks through all cards until it finds
the card with target name, there is a chance when two clients are to
open the lessee card at the same time using drmOpen, card0 is opened
temporarily by the first drmOpen when searching for the lessee card,
and fail the second drmOpen as card0 is detected as opened in master
mode.

This change relaxes the check that when card0 is opened as master
and if lessee card tries to open again, the lessee card open will
force the current card0 master to drop the master and set the
lessee card as the new master.

Change-Id: I342d4462e83beddef138f8e13986b65ac54b1640
Signed-off-by: Xiaowen Wu <wxiaowen@codeaurora.org>