disp: msm: fix build issue when using common defconfig
It will encounter compile error after consolidata the
defconfig of all Gen3 platform into a common deconfig.
The root cause is that the new defconfig include more
than one platform macro definition(CONFIG_ARCH_*), the
judgment in the Makefile fails, augen3dispconf will not
be included. The repair method is to use findstring to
replace filter in the judgment of Makefile.

Change-Id: I3026701fa8024a8af66a5613012d4b1d41c6c0b3
Signed-off-by: Zhiqiang Liu <zhiqliu@codeaurora.org>