disp: msm: sde: clear num_roi_misrs when allocating misr roi
Encoder's mode_set is called when mode is updated, and disable is
called only when CRTC is disabled. In the case of resume, mode_set
is firstly called when CRTC is disabled, and then called again
when CRTC is enabled. num_roi_misrs need to be cleared in the 2nd
call otherwise it will result invalid roi misr IRQ.

Change-Id: I9d619011551cb157e44b92905f35837f18603387
Signed-off-by: Xiaowen Wu <wxiaowen@codeaurora.org>