disp: msm: dp: check new crtc active state for bond mode disable
When bond mode is disabled, there are cases when CRTC is still
attached but active is set to 0. In current check if CRTC is
attached, we'll check connector's connect state and if one is
not connected, atomic check will fail. This change will check
for both CRTC attach state and active state.

Change-Id: I40d0a1134ad2f348305f04b231008b6e636a5aae
Signed-off-by: Xiaowen Wu <wxiaowen@codeaurora.org>