disp: msm: dsi: add properties for PHY lane mapping
Add DSI PHY device tree property "qcom,platform-lane-map-v2"
to specify the PHY lane mapping. It have the same definition of
the same name property in the DSI display device tree.

If this property is not defined, lane mapping configuration from
the DSI display will be used. Otherwise, the configuration defined in
PHY node will override the one from the display.

This is useful when bonding two DSIs as a single display, but have
different lane mapping configuration.

Change-Id: Ic3acffde61031724dc19daca6155dc6efc0a2075
Signed-off-by: Jim Wang <jimwan@codeaurora.org>