disp: msm: dp: add PLUG interrupt support in lphw HPD handler
Currently driver relies on host_deinit to re-enable PLUG detection,
which is controlled from user space. In the case if host_deinit is
late or not called, we'll miss the PLUG interrupt. This change adds
PLUG interrupt support in lphw HPD handler, so HPD can be detected
before user calls to host_deinit.

Change-Id: I8992feb7fa0055e61fce1914095110dc36d80c1d
Signed-off-by: Xiaowen Wu <wxiaowen@codeaurora.org>