disp: msm: shd: fix misr hardware signature error issue
The signature values are mismatched with the software calculated
results in shared display case. The root cause is that misr
hardware configuration, dsc enable bit and hardware control flush
bit are not written to the related registers in the specified order.
So cached the misr hardware configuration and dsc config in shared
display and write all the values to registers at hardware flush stage.

Change-Id: I0f69479402c82befee2e777d0b5aea258206fc93
Signed-off-by: Zhiqiang Liu <zhiqliu@codeaurora.org>