disp: msm: dp: reset edid segment and address when reading port edid
Currently, edid read failures are seen when transitioning from a
resolution with larger edid size to a resolution with smaller
edid size. Edid segment and address values are not getting reset,
and any edid read after switching to a lower resolution
reads the incorrect segment.

This change clears edid segment and resets edid address in the
read_edid function so that the affected values are cleared every
time edid is read.

Change-Id: Iedb231c877b85d645ca74071f9c663393a204ff7
Signed-off-by: Sudarsan Ramesh <sudarame@codeaurora.org>
Signed-off-by: Xiaowen Wu <wxiaowen@codeaurora.org>